+++ banyand/internal/storage/storage.go
@@ -109,6 +109,8 @@ type TSDB[T TSTable, O any] interface {
 	Tick(ts int64)
 	UpdateOptions(opts *commonv1.ResourceOpts)
 	TakeFileSnapshot(dst string) error
+	GetExpiredSegmentsTimeRange() *timestamp.TimeRange
+	DeleteExpiredSegments(timeRange timestamp.TimeRange) int64
 }
 
 // Segment is a time range of data.